requirements.txt
qdrant-client>=1.9.1
sentence-transformers>=3.0.0
torch
transformers
beir~=2.0
pandas
tqdm
python-dotenv
streamlit
rank-bm25>=0.2.2
